refactor(website): change Github to GitHub
